Climate variables calculation

bFFP : Day of the year on which the Frost-Free Period begins

eFFP : Day of the year on which the Frost-Free Period ends

Based on simplified Penman - Monteith method from Hogg (1997)

Usage

calc_DD_below_0(m, tm)

calc_DD_above_5(m, tm, region)

calc_DD_below_18(m, tm)

calc_DD_above_18(m, tm, region)

calc_EMT(t_min_list, mcmt, td)

calc_EXT(t_max_list, mwmt, td)

calc_MCMT(t_ave_list)

calc_MWMT(t_ave_list)

calc_TD(mcmt, mwmt)

calc_Eref(m, tmmin, tmmax, latitude)

calc_CMD(Eref, PPT)

calc_bFFP(td, NFFD, t_min_list)

calc_eFFP(NFFD, t_min_list)

calc_FFP(bFFP, eFFP)

calc_NFFD(m, tmin)

calc_PAS(m, tave, PPT)

calc_RH(tmmin, tmmax)

calc_PET(tave, tmmin, tmmax, alt)

calc_SVP(t)

Arguments

m

integer. Month of the year.

tm

numeric. Monthly mean temperature for the m month.

region

character. One of either "All", "West", "East".

t_min_list

list of monthly minimum temperature for each month in order.

mcmt

Mean Temperature of the Coldest Month

td

numeric. Difference between the mean warmest monthly temperature and the mean coldest monthly temperature.

t_max_list

list of monthly maximum temperature for each month in order

mwmt

Mean Temperature of the Warmest Month

tmmin

numeric. Monthly mean minimum air temperature.

tmmax

numeric. Monthly mean maximum air temperature.

latitude

numeric. Point latitudes.

Eref

Reference evaporation

PPT

numeric. Precipitation in mm.

NFFD

numeric. Number of Frost-Free Days.

bFFP

numeric. Day of the year on which the Frost-Free Period begins

eFFP

numeric. Day of the year on which the Frost-Free Period ends

tmin

numeric. Minimum temperature for the m month.

tave

numeric. Monthly average air temperature.

alt

numeric. Altitude in m.

t

numeric. Air temperature in degree Celsius.

Value

numeric. Reference evaporation (Eref)

numeric. Climatic moisture deficit

numeric. Julian day on which the Frost-Free Period begins

numeric. Julian day on which the Frost-Free Period ends

numeric. Frost-Free Period in days

numeric. Number of Frost Free Days

numeric. Precipitation As Snow

numeric. Relative Humidity

References

Hogg, E.H. (1997). Temporal scaling of moisture and the forest-grassland boundary in western Canada. Agricultural and Forest Meteorology, Research on Forest Environmental Influences in a Changing World, 84, 115–122.

Examples

if (FALSE) { # \dontrun{
climr:::calc_DD_below_0(2, -14)
} # }
if (FALSE) { # \dontrun{
climr:::calc_DD_above_5(2, -14, "All")
} # }
if (FALSE) { # \dontrun{
climr:::calc_DD_below_18(2, -14)
} # }
if (FALSE) { # \dontrun{
climr:::calc_DD_above_18(2, -14, "All")
} # }
if (FALSE) { # \dontrun{
t_min_list <- list(
  "1" = -35, "2" = -32, "3" = -25, "4" = -10,
  "5" = -5, "6" = 3, "7" = 15, "8" = 17, "9" = 10, "10" = -5,
  "11" = -20, "12" = -30
)

climr::calc_bFFP(td = 30, NFFD = 10, t_min_list = t_min_list)
} # }
if (FALSE) { # \dontrun{
t_min_list <- list(
  "1" = -35, "2" = -32, "3" = -25, "4" = -10,
  "5" = -5, "6" = 3, "7" = 15, "8" = 17, "9" = 10, "10" = -5,
  "11" = -20, "12" = -30
)

climr::calc_eFFP(NFFD = 10, t_min_list = t_min_list)
} # }
if (FALSE) { # \dontrun{
climr::calc_FFP(bFFP = 214.5964, eFFP = 265.4581)
} # }
if (FALSE) { # \dontrun{
climr::calc_NFFD(3, 2.05)
} # }
if (FALSE) { # \dontrun{
climr::calc_PAS(4, 2, 600)
} # }
if (FALSE) { # \dontrun{
climr::calc_RH(tmmin = 10, tmmax = 40)
} # }