climr: Downscaling of global climate data

This package replicates the basic approach of ClimateNA (climatena.ca), but with (1) improved speed, (2) multiple runs of each GCM, (3) cloud delivery, and (4) programmatic access via R. A simple downscaling method is used to bias-correct low-resolution (~100km) global climate model anomalies to high-resolution (2km) maps of historical climate, with further elevation adjustment to user-specified elevation grids/points based on empirical lapse rates (local relationship of climate to elevation) of the 2km climate maps. Elevation-adjusted monthly values of basic climate elements (temperature and precipitation) are then used to estimate derived variables (e.g., degree-days) based on published equations and parameters from Wang et al. 2016.

Author

Maintainer: Colin Mahony Colin.Mahony@gov.bc.ca (ORCID)

Authors:

Kiri Daust kiri.daust@gov.bc.ca

Bruno Tremblay bruno@boostao.ca (ORCID)

Ceres Barros ceres.barros@gov.bc.ca (ORCID)

Other contributors:

Francois Bornais francois@boostao.ca [contributor]

Province of British Columbia [copyright holder, funder]