List available runs, global climate models (GCMs), time periods and scenarios (SSPs)

list_gcms lists available global climate models.

list_ssps lists available greenhouse gas concentration scenarios (SSP-RCPs).

list_gcm_periods lists available 20-year normal periods for GCM simulations.

list_runs_ssp lists available runs for a given GCM/ssp.

list_runs_historic lists available runs from the historical simulation (1851-2014) for a specified GCM.

list_refmaps lists available reference maps of gridded climate normals

list_obs_periods lists available normal periods for observational climate data

list_vars lists available climate variables

list_obs_years lists available years for time series of observational climate data

list_gcm_ssp_years lists available years for time series of global climate model future simulations

list_gcm_hist_years lists available years for time series of global climate model historical simulations

Usage

list_gcms()

list_ssps()

list_gcm_periods()

list_runs_ssp(gcm, ssp)

list_runs_historic(gcm)

list_refmaps()

list_obs_periods()

list_vars(set = c("All", "Monthly", "Seasonal", "Annual"), only_extra = FALSE)

list_obs_years()

list_gcm_ssp_years()

list_gcm_hist_years()

Arguments

gcm

Name of GCM

ssp

Name of scenario Must be one of the elements in list_ssps().

set

character. One of c("All", "Monthly", "Seasonal", "Annual") or any combination thereof. Defaults to "All".

only_extra

logical. Should Tmin, Tmax and PPT be excluded? Defaults to FALSE.

Value

a character vector.

Details

The eight-gcm ensemble recommended by Mahony et al. (2022) can be specified as gcms = list_gcm()[c(1,4,5,6,7,10,11,12)].

Currently available reference maps of gridded climate normals (list_refmaps()) are:

"refmap_climatena" for Climate NA derived normals

"refmap_climr" for a composite of BC PRISM, adjusted US PRISM and DAYMET (Alberta and Saskatchewan), covering North America

A lookup table of variable names, units, types, and other attributes is available by calling data(variables).