Bilinear interpolation point extraction from raster bands

Bilinear interpolation point extraction from raster bands

Usage

extract_db(
  rastertbl,
  layers = data.table::data.table(var_nm = character(), laynum = integer()),
  VAR = NULL,
  hull = NULL
)

Arguments

rastertbl

The name of the raster table to extract from.

layers

A data.table with column var_nm (names of the bands) and

VAR

In case of rasters split across different tables. One VAR in each. The first VAR will be used for replacement in labels names. Assuming all tables have the same structure, only different VAR in them.

hull

An optional hull well-known text geometry to prefilter raster table before merging points projected in srid 4326. laynum (index number of the bands to extract).

Value

A data.frame one row per point id in xyz with an ID column + one columns for each bands in rastertbl name nm.