Function for plotting time series for gcms or compiled ensemble

Function for plotting time series for gcms or compiled ensemble

Usage

plot_preprocess_ensemble(
  x,
  var,
  scenarios.selected,
  scenarios,
  showrange = TRUE,
  simplify = TRUE,
  gcm = NULL,
  pal,
  pal.scenario,
  pal.gcms,
  refline = FALSE,
  showmean = TRUE,
  endlabel = "change",
  element,
  compile = TRUE,
  var2 = NULL,
  element1,
  element2,
  yeartime.names,
  yeartimes,
  yeartime
)

Arguments

x

climate data

var

TODO

scenarios.selected

TODO

scenarios

TODO

showrange

logical. Plot a shaded region indicating the minimum and maximum of the selected ensemble of GCM simulations for each selected scenario.

simplify

logical. Simplify the ensemble range and mean using a smoothing spline.

gcm

TODO

pal

character. color palette. Options are "scenario", for use when comparing scenarios, and "gcms", for use when comparing GCMs.

pal.scenario

TODO

pal.gcms

TODO

refline

logical. Plot the 1961-1990 reference period mean for the selected var and extend this line to the year 2100 as a visual reference.

showmean

logical. Plot the ensemble mean time series. Multi-model ensemble means are calculated from the mean of simulations for each model.

endlabel

character. Add a label to the end of each simulated time series. Options are "change", to indicate the change in year 2100 relative to the 1961-1990 baseline, or "gcms" to indicate the global climate model.

element

TODO

compile

logical. Compile multiple global climate models into a multi-model ensemble. If FALSE the single-model ensembles are plotted individually.

var2

character. A second climate var to plot in combination with var1. options are list_vars().

element1

TODO

element2

TODO

yeartime.names

TODO

yeartimes

TODO

yeartime

TODO