Download historic period anomalies.

historic_input produces anomalies of historic observed climate for a given period.

historic_input_ts produces anomalies of historic observed climate for a given time series.

Usage

historic_input(dbCon, bbox = NULL, period = list_historic(), cache = TRUE)

historic_input_ts(dbCon, bbox = NULL, years = 2010:2022, cache = TRUE)

Arguments

dbCon

A db connection object created by data_connect.

bbox

numeric. Vector of length 4 giving bounding box of study region, created by get_bb().

period

character. Vector of labels of the periods to use. Can be obtained from list_historic(). Default to "2001_2020".

cache

logical. Specifying whether to cache new data locally or no. Defaults to TRUE.

years

numeric. Years to retrieve historic anomalies for, in 1902:2022. Default 2010:2022

Value

A list of SpatRasters, each with possibly multiple layers, that can be used with downscale(). Each element of the list corresponds to a particular period, and the values of the SpatRaster are anomalies of the historic period compare to the normal period.

List of length 1 containing a SpatRaster

Details

Generally, this function should only be used in combination with downscale() as the values returned in the rasters are anomalies compared to the 1961-1990 normal period, and are usually not meaningful without the whole downscale process.

The returned raster contains anomalies for each year specified in years. In general this function should only be used in conjunction with downscale().

See also